+++ Core/Session/ExperienceProvider.cs
@@ -5,11 +5,11 @@ namespace Core.Session
 {
     public static class ExperienceProvider
     {
-        public static double[] GetExperienceList()
+        public static float[] GetExperienceList()
         {
             var dataConfig = new DataConfig();
             var json = File.ReadAllText($"{dataConfig.Experience}exp_tbc.json");
-            var expList = JsonConvert.DeserializeObject<double[]>(json);
+            var expList = JsonConvert.DeserializeObject<float[]>(json);
             return expList;
         }
     }
